feat!: add measure_leaked, add bool/uint future results, add simple leakage error model (#25)
This reflects a change to the helios QIS, as well as the interface to the runtime and error model interfaces. It enables calling measure_leaked, returning 0 for false, 1 for true, and 2 for leaked. This required adding bool and uint futures rather than assuming all futures are boolean. It also required a new error model for leakage in order to test the pipes.