Fix: Accept library paths from component plugins before invocation (#43)
In the python interface to runtime plugin specifications (Simulator,
Error Model, Runtime), there is a `library_search_dirs` property.

The intent is that, when these are provided, the selene frontend should
add the resulting list of paths to the PATH, LD_LIBRARY_PATH or
DYLD_LIBRARY_PATH environment variables so that the dynamic linker can
find dependencies.

Unfortunately, this parameter is not currently being utilised by selene.
This remedies that.