feat: Fine-grained timeout configuration (#69)
The timeout parameter for run_shots required more fine-grained control
so that different timeout aspects can be configured.

Previously:
- the `timeout` parameter was a `timedelta|None` that corresponded to a
socket read timeout, which was counter-intuitive for many users.

With this PR:
- The `timeout` parameter is a `Timeout|timedelta|None`.
- `Timeout` is a new class that allows independent setting of:
  - Overall simulation time: maximum time for all shots to finish
  - Per-shot simulation time: maximum time for a shot to finish
- Per-result simulation time: maximum time to wait between results in a
shot
- Backend startup time: maximum time for the frontend TCP host to wait
for client connections before giving up
- If given a timedelta, this sets the overall time.

As the original input format (a datetime) is still supported, and will
have the behaviour that users are expecting anyway, that aspect is not
necessarily breaking.

This PR also introduces a SeleneTimeoutError. This is triggered on any
timeout _except_ for the backend startup time, as this keeps the
existing behaviour of SeleneStartupError.