fix: correct post_runtime duration metric (#74)
This metric used to add the duration to its previous value. This lead to
some issues with overlapping batches so it was amended to use a maximum
of itself and the end time of the batch. However, this change was faulty
and added the aforementioned maximum to itself rather than setting its
value as that maximum. This PR fixes that.