chore: Bump tket version; add wasm + gpu to the hugr-qis registry (#80)
Note that this makes `selene-hugr-qis-compiler` point to a guppy branch
`accept-any-selene`. This is a version of guppy which generates the
appropriate compiled code, but loosens the `selene-hugr-qis-compiler`
constraint so it doesn't point to this branch (unlike in current guppy
`main`). This allows `uv` to use the local version for testing, without
pulling this branch in as a transitive dependency and causing a
conflict.

Note: guppylang had a higher `tqdm` dep than was being pulled in before,
so I've raised it to match and fixed the errors in
https://github.com/CQCL/selene/pull/80/commits/22492ac807ffd0993c8ee24fa36d5c98f1759dd5

### Release plan
* Merge this PR and make a release of `selene-hugr-qis-compiler`
* Update `guppylang` to use the released version
* Release `guppylang`
* At which point, we can point `selene-hugr-qis-compiler` to the
released `guppylang` version

---------

Co-authored-by: Jake Arkinstall <65358059+jake-arkinstall@users.noreply.github.com>
Co-authored-by: Seyon Sivarajah <seyon.sivarajah@quantinuum.com>