Feat: correct shot end strategy and error processing (#91)
Currently metadata is being written *before* shot_end is called, which
means that teardown information (e.g. instruction logs) are not
preserved. This PR fixes that.

This highlights that the emission of extra information after shot_end is
quite useful, while in the current implementation we prevent output upon
a panic. As such, this PR also modifies error handling to allow
additional data to be provided after the exception has been introduced,
allowing for inspection of metadata (e.g. instruction logs) even after
the panic message has been emitted to the results stream. It does this
by removing an inherent assumption that the last 4 entries correspond
directly to the error details, and upon missing information it now
returns a useful error message instead of None.

The latter topic currently only makes sense in the context of
parse_results=False, as the default (parse_results=True) behaviour is to
raise an exception upon spotting the panic message in the results
stream, thus eliminating the extraction of previous results.