feat: Result stream handling refactor (#93)
Problems and their solutions in this merge:

- We expected a successful connection to be immediately started with a
shot, which means any error that happened between load and shot start
could be designated a shot error.
- Fixed by removing SHOT_BOUNDARY in favour of SHOT_START and SHOT_END
control messages, and restructuring shot result extraction code to
differentiate between in-shot and out-of-shot messages.

- One example of such a failure is when a SeleneInstance has
successfully parsed a configuration file, and has launched the TCP
client for results sending, but later returns an Err because a plugin
failed to load. Leaving the scope of the function means that the output
stream closes, and we don't have an opportunity to write the panic to
the stream.
- Fixed by moving the loading of plugins to within the Emulator class,
making it a single Result call from the SeleneInstance constructor, then
printing a panic to the stream _within_ the SeleneInstance constructor
before returning in case of error.

- Processes assume they can write to the results stream, leading to
possible panics via unwrap() upon early termination of the frontend
(e.g. when another process has printed a panic).
  - Fixed by adding a fallible_print_panic method to SeleneInstance.

- Too many assumptions were being made about the types coming in from
the result stream.
- Fixed by validating early on and encoding values in a richer set of
classes depending on their intent.

- The 'unparsed' and 'parsed' interfaces were too different, with
separate implementations, leading to potential divergence.
- Fixed by restructuring the data flow in a manner that takes into
account the requirement for distinct exception-based and stream-based
flows in one straightforward manner, arriving at (IMO) a much clearer
result parsing codebase.