feat: Add support for object files provided as bytes (#94)
This PR adds support for a requested feature: accepting `bytes` objects
to `selene_sim.build()` that represent the contents of object files - we
have support for LLVM bitcode through the same fashion (bytes or the
file).

It also performs more validation on bitcode bytes, as the previous
approach was simply to check that it was a bytes object with required
symbols present. The magic number at the start of a bitcode file is now
verified.

Finally, we perform additional build testing. Upon each build test, we
run through intermediate artifacts, convert them to their 'in memory'
equivalent and build against those too. This is all done in strict build
mode, such that the output of every build step is validated against its
Kind (more expensive) instead of assuming validity (faster, default).