chore: Manage platform tags better in hatch (#95)
Currently we manage platform tags by generating the wheel with
default-inferred tags, then renaming the wheel.

This is because if we let hatch infer the target directly with
`infer_tag`, it uses the current interpreter and ABI (cp310-cp310) as a
constraint for the wheel. This is not what we want - selene compiled
libraries don't interact with python itself (outside of popen calls).

The current approach allows hatch to infer the tag, then renames
`cp310-cp310` to `py3-none`. Unfortunately this is incomplete - the
WHEEL file within still has cp310 inside, and this confuses package
managers. `uv sync` on a workspace with selene-sim as a dependency will
soft-check selene-sim every time because of the mismatch. This
inconvenience is minor but could lead to much bigger problems with other
UV versions or other package managers - it needs to go.

As such, I here provide a different approach. Use hatch's approach to
identifying the right platform tag, but only take the platform part of
it, and hardcode 'py3-none-' as intended.