chore: Add include files to source tree, validate them in CI instead of generating them in CI (#97)
Closes #92.

This avoids the requirement for cargo-expand and cbindgen in the build
of a package itself - regenerating bindings can be done with `just
generate-bindings`. This is validated in CI by running `just
generate-bindings` and asserting that no diff is produced.

This simplifies the build for users, makes it clearer when include files
change (as their diffs will be present in commits), and should speed up
CI.