feat: Cleanup error'd processes before log collection (#98)
As soon as an error message is received through the results stream, we
read stdout and stderr into the exception message.

There is a race condition whereby logs are collected before the process
has flushed to stdout and stderr. This condition revealed itself in
https://github.com/CQCL/selene/actions/runs/19111445555/attempts/1.

This is resolved in two steps:
1. Print a startup error _before_ notifying the results stream. 
2. Passing the process handle itself to the result handling, we can
terminate the process _before_ gathering logs.

The latter requires us to make SeleneProcess importable from
result_handling, so it has been moved to a new file. To retain the
existing interface, it is re-exported from SeleneInstance, such that
`from selene_sim.instance import SeleneProcess` is still valid.

Driveby:
- Strip EXIT:INT: from error messages (left from previous merge)
- correct data stream logging - close self.logfile_handle